Se añaden cambios de tipo de dato en el campo flete, tabla historial y tambien se añade un nuevo archivo dataset

+++ database/migrations/2008_06_16_031928_create_historiales_table.php
@@ -23,7 +23,7 @@ class CreateHistorialesTable extends Migration
             $table->string('origen')->nullable();
             $table->string('destino')->nullable();
             $table->string('cliente')->nullable();
-            $table->integer('flete')->nullable();
+            $table->float('flete')->nullable();
             $table->timestamps();
         });
     }
